feat: add basic example
