+++ apps.toml
@@ -1,0 +1,12 @@
+# 利用文件描述 依赖关系，然后动态的选择是否引入模块
+# 如果合并成一个文件，需要设置继承资源
+# 如果存在多个程序都直接依赖一个模块且这个模块拥有自己独特的资源，那么这个模块无法被合并
+[[tasks]]
+name = "uart-thread"
+file = "uart-thread.elf"
+mem = ["VIRTIO_MMIO_VIRT_ADDR", "PL011_ADDR", "0x1000"]
+
+[[tasks]]
+deps = ["uart-thread"]
+name = "test-demo"
+file = "test-demo.elf"
