+++ crates/srv-iface/Cargo.toml
@@ -3,8 +3,13 @@ name = "srv-iface"
 version = "0.1.0"
 edition.workspace = true
 
+[features]
+uart-ipc = []
+
 [dependencies]
+sel4 = { workspace = true }
 common = { workspace = true }
 linkme = { workspace = true }
 spin = { workspace = true }
 num_enum = { version = "0.7.3", default-features = false }
+zerocopy = { workspace = true }
