+++ tasks/test-demo/Cargo.toml
@@ -3,6 +3,11 @@ name = "test-demo"
 version = "0.1.0"
 edition.workspace = true
 
+[features]
+uart-ipc = ["srv-iface/uart-ipc"]
+
+default = ["uart-ipc"]
+
 [dependencies]
 log = "0.4"
 config = { workspace = true }
